+++ app/src/main/java/dev/esnault/bunpyro/data/repository/settings/ISettingsRepository.kt
@@ -1,6 +1,7 @@
 package dev.esnault.bunpyro.data.repository.settings
 
 import dev.esnault.bunpyro.domain.entities.grammar.AllGrammarFilter
+import dev.esnault.bunpyro.domain.entities.settings.ExampleDetailsSetting
 import dev.esnault.bunpyro.domain.entities.settings.FuriganaSetting
 import dev.esnault.bunpyro.domain.entities.settings.NightModeSetting
 
@@ -10,6 +11,8 @@ interface ISettingsRepository {
 
     suspend fun getFurigana(): FuriganaSetting
     suspend fun setFurigana(setting: FuriganaSetting)
+
+    suspend fun getExampleDetails(): ExampleDetailsSetting
 
     suspend fun setAllGrammarFilter(filter: AllGrammarFilter)
     suspend fun getAllGrammarFilter(): AllGrammarFilter